Add bash scripts to build, deploy and run all EPC elements for testing.
The top level bash script 'run.sh' includes several commands to make
it easy to (re)build individual or all centos8 docker containers.

    ./run.sh [build | run | logs | configs | pcap | stop]

To build containers:

    ./run.sh build [ cassandra | hss | mme | spgwc | spgwu-tiny ]

If an individual container is specified, only that container will
be built.  Otherwise all of the containers will be built.

To run the EPC:

    ./run.sh start

This will deploy and configure all of the containers, finally running
the binary for the network element.

To get the log files:

    ./run.sh logs [ hss | mme | spgwc | spgwu-tiny ]

This will copy the log files in to the 'archives' subdirectory and
show them on the terminal.

If an individual container is specified, only that container will be
copied and shown.  Otherwise all of the containers will be copied.

To get the pcap files:

    ./run.sh pcap [ hss | mme | spgwc | spgwu-tiny ]

This will copy the pcap files in to the 'archives' subdirectory and
show them on the terminal.

If an individual container is specified, only that container will
be copied.  Otherwise all of the containers will be copied.

To stop the EPC:

    ./run.sh stop

This will stop all of the containers and remove the docker networks.