Add new docker_control scripts to build, deploy and run all EPC elements.
Add a new bash script './scripts/docker_control.sh' to simplify building,
deploying and running EPC network element containers with Docker.

A new bash script './scripts/docker_control.sh' includes several commands to make
it easy to (re)build, run, and control individual or all centos8 docker containers.

    ./scripts/docker_control.sh [start | stop | build | logs | configs | pcap]

To build containers:

    ./scripts/docker_control.sh build [ cassandra | hss | mme | spgwc | spgwu-tiny ]

If an individual container is specified, only that container will
be built.  Otherwise all of the containers will be built.

To run the EPC:

    ./scripts/docker_control.sh start

This will deploy and configure all of the containers, finally running
the binary for the network element.

To get the log files:

    ./scripts/docker_control.sh logs [ hss | mme | spgwc | spgwu-tiny ]

This will copy the log files in to the 'archives' subdirectory and
show them on the terminal.

If an individual container is specified, only that container will be
copied and shown.  Otherwise all of the containers will be copied.

To get the pcap files:

    ./scripts/docker_control.sh pcap [ hss | mme | spgwc | spgwu-tiny ]

This will copy the pcap files in to the 'archives' subdirectory and
show them on the terminal.

If an individual container is specified, only that container will
be copied.  Otherwise all of the containers will be copied.

To stop the EPC:

    ./scripts/docker_control.sh stop

This will stop all of the containers and remove the docker networks.