Fix bug with configuring mme WWR with 2 digit MNC values.
Fix a bug in generate_mme_config_script.py where the eDNS address
generated for the SPGWC (WRR_LIST_SELECTION) would not add the leading
zero to the MNC value when the MNC is only two digits.

According to TS 23.003 section 19:

The Home Network Realm/Domain shall be in the form of
"epc.mnc<MNC>.mcc<MCC>.3gppnetwork.org", where "<MNC>" and "<MCC>"
fields correspond to the MNC and MCC of the operator"s PLMN. Both the
"<MNC>" and "<MCC>" fields are 3 digits long. If the MNC of the PLMN
is 2 digits, then a zero shall be added at the beginning.