Add new EPC docker tutorial EPC-in-a-box.
Add new EPC docker tutorial EPC-in-a-box which documents the changes
that have been made to openairinterface5g to add vagrant virtual
machines to and the new 'docker_control' scripts added to OAI-EPC
to support a full 4G network simulation which can be run on a single
machine.

The EPC-in-a-box document unifies the desparate and often outdated
documentation from OAI-EPC and openairinterface5g into a simple
step-by-step proceedure to create a containerized EPC using the new
'docker_control' scripts and test it with a simulated eNodeB and UE
in virtual machines.